requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-genai
chromadb
pandas
openpyxl

